refactor: change hyphens file names to camelCase
