+++ events/onStartUp.js
@@ -1,4 +1,4 @@
-import { logCurrentPath } from "../helpers/log-current-path.js";
+import { logCurrentPath } from "../helpers/logCurrentPath.js";
 
 export function onStartUp(username) {
   console.log(`Welcome to the File Manager, ${username}`);
