+++ helpers/checkIfPathExists.js
