+++ helpers/logCurrentPath.js
