--- helpers/parseArguments.js
@@ -1,3 +1,0 @@
-export function parseArguments(command) {
-  command.match()
-}
