+++ operations/add.js
@@ -1,7 +1,7 @@
 import { Errors } from "../constants/errors.js";
 import fs from 'fs';
-import { checkIfPathExists } from "../helpers/check-if-path-exists.js";
 import { checkArgumentsAmount } from "../helpers/checkArgumentsAmount.js";
+import { checkIfPathExists } from "../helpers/checkIfPathExists.js";
 
 export async function add({ args }) {
   if (!checkArgumentsAmount(args.length, 1)) return;
