+++ operations/cp.js
@@ -1,7 +1,7 @@
 import path from "path";
 import fs from "fs";
 import { Errors } from "../constants/errors.js";
-import { checkIfPathExists } from "../helpers/check-if-path-exists.js";
+import { checkIfPathExists } from "../helpers/checkIfPathExists.js";
 import { toAbsoultePath } from "../helpers/toAbsolutePath.js";
 import { checkArgumentsAmount } from "../helpers/checkArgumentsAmount.js";
