+++ operations/rn.js
@@ -1,6 +1,6 @@
 import fs from "fs/promises";
 import { Errors } from "../constants/errors.js";
-import { checkIfPathExists } from "../helpers/check-if-path-exists.js";
+import { checkIfPathExists } from "../helpers/checkIfPathExists.js";
 import path from "path";
 import { toAbsoultePath } from "../helpers/toAbsolutePath.js";
 import { checkArgumentsAmount } from "../helpers/checkArgumentsAmount.js";
